+++ docs/Whats-New.md
@@ -260,6 +260,7 @@ New:
 - Script actions for manipulating script flows with conditional jumps (by FS-21)
 - Customizable ROF random delay (by Starkku)
 - Animation with `Tiled=yes` now supports `CustomPalette` (by ststl)
+- Script action for repairing destroyed bridges (by FS-21)
 
 Vanilla fixes:
 - Allow AI to repair structures built from base nodes/trigger action 125/SW delivery in single player missions (by Trsdy)